dev_tools/requirements.txt
torch>=1.12.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
pyyaml>=6.0
jinja2>=3.1.0
pygments>=2.11.0
markdown>=3.3.0
psutil>=5.8.0
gputil>=1.4.0
astor>=0.8.1